implement positional struct tag parsing add positional tests improve documentation improve document loading
