+++ docs/_sass/custom/custom.scss
@@ -1,3 +1,4 @@
+
 pre {
   font-size: 1.2rem;
 }
