+++ docs/guides/advanced-features.md
@@ -2,12 +2,12 @@
 layout: default
 title: Advanced Features
 parent: Guides
-nav_order: 4
+nav_order: 5
 ---
 
 # Advanced Features
 
-## Struct Tag formats
+## Struct Tag formats (New vs Old)
 
 | Feature | New Format | Old Format (Deprecated) |
 |---------|------------|------------------------|
@@ -26,6 +26,7 @@ nav_order: 4
 | SliceCapacity | capacity:value | N/A |
 | Accepted Values | accepted:{pattern:json\|yaml,desc:Format type},{pattern:text\|binary,desc:Type} | accepted:pattern:json\|yaml,desc:Format type |
 | Dependencies | depends:{flag:output,values:[json,yaml]},{flag:mode,values:[text]} | depends:flag:output,values:[json,yaml] |
+| Positional Flag | pos:{at:start\|end,idx:value} | N/A |
 
 The new format offers several advantages:
 - Namespace Isolation: Using goopt: prefix prevents conflicts with other tag parsers
@@ -139,3 +140,149 @@ Important notes:
 3. Memory safety is ensured by flag registration - only valid paths are accepted
 4. Attempting to use an index outside the registered range results in "unknown flag" error
 5. Slice bounds are tracked for user feedback when using `NewParserFromStruct` but are not required for memory safety
+
+## Advanced Flag features
+
+
+### Flag Types
+Flag types are defined in the `github.com/napalu/goopt/types` package and are used to define the type of a flag.
+
+The following types are supported:
+- `Single`: A flag which expects a single value - if you do not specify a type, this is the default
+- `Standalone`: A boolean flag which is true by default - can be overriden with a false value on the command line
+- `Chained`: A flag which is evaluated as a list of values - by default, the delimiters are ',' || r == '|' || r == ' '. This can be overridden by providing a custom delimiter function `ListDelimiterFunc`.
+- `File`: A flag which expects a file path and is set to the contents of the file
+
+The flag type can be specified using the `WithType` option function. The flag type can also be specified in the struct tag using the `type` tag.
+Flag types in structs which are parsed from a struct (see for example `NewParserFromStruct`) are automatically inferred from the field type.
+
+### Secure Flags
+Some flags contain sensitive information (like passwords) and should be kept secure during input. `goopt` supports secure flags, which prompt the user without echoing input back to the terminal.
+
+```go
+package main
+
+import (
+	"os"
+	"fmt"
+	"github.com/napalu/goopt"
+    "github.com/napalu/goopt/types"
+)
+
+func main() {
+	parser := goopt.NewParser()
+
+	// Define a secure flag
+	parser.AddFlag("password", goopt.NewArgument("p", "password for app", types.Single, true, goopt.Secure{IsSecure: true, Prompt: "Password: "}, ""))
+
+	// Parse the arguments
+	if parser.Parse(os.Args) {
+		password, _ := parser.Get("password")
+		fmt.Println("Password received (but not echoed):", password)
+	} else {
+		parser.PrintUsage(os.Stdout)
+	}
+}
+```
+
+Secure flags ensure that user input for sensitive fields is hidden, and can optionally display a prompt message.
+
+### Dependency Validation
+
+`goopt` allows you to define dependencies between flags, ensuring that certain flags are present or have specific values when others are set. This is useful for enforcing consistency in user input.
+
+```go
+package main
+
+import (
+	"os"
+	g "github.com/napalu/goopt"
+    "github.com/napalu/goopt/types"
+)
+
+func main() {
+    parser := g.NewParser()
+    
+    // Define flags
+    parser.AddFlag("notify", g.NewArg( 
+        g.WithDescription("Enable email notifications"), 
+        g.WithType(types.Standalone))
+    parser.AddFlag("email", g.NewArg(
+        g.WithDescription("Email address for notifications"), 
+        g.WithType(types.Single))
+    
+    // Set flag dependencies - new style
+    parser.AddDependencyValue("email", "notify", []string{"true"})
+    
+    // Or using WithDependencyMap in flag definition
+    parser.AddFlag("email", g.NewArg(
+        g.WithDescription("Email address for notifications"),
+        g.WithType(types.Single),
+        g.WithDependencyMap(map[string][]string{
+            "notify": {"true"},
+        })))
+    
+    // Parse the arguments
+    if !parser.Parse(os.Args) {
+        parser.PrintUsage(os.Stdout)
+    } else {
+        email, _ := parser.Get("email")
+        fmt.Println("Email notifications enabled for:", email)
+    }
+}
+```
+
+In this example, the email flag is only valid if the notify flag is set to true. If the dependency is not satisfied, goopt will add a warning which can be displayed to the user (see `goopt.GetWarnings()`).
+
+### Command-Specific Flags
+
+`goopt` supports associating flags with specific commands or subcommands. This allows you to define different behaviors and options for different parts of your application.
+
+```go
+package main
+
+import (
+	"os"
+	"github.com/napalu/goopt"
+    "github.com/napalu/goopt/types"
+)
+
+func main() {
+	parser := goopt.NewParser()
+
+	// Define commands
+	parser.AddCommand(&goopt.Command{
+		Name: "create",
+		Subcommands: []goopt.Command{
+			{Name: "user"},
+			{Name: "group"},
+		},
+	})
+
+	// Define flags for specific commands
+	parser.AddFlag("username", goopt.NewArgument("Username for user creation", types.Single), "create user")
+	parser.AddFlag("email", goopt.NewArgument("Email address for user creation", types.Single), "create user")
+
+	// Parse the command-line arguments
+	if parser.Parse(os.Args) {
+		username, _ := parser.Get("username")
+		email, _ := parser.Get("email")
+		fmt.Println("Creating user with username:", username)
+		fmt.Println("Email address:", email)
+	}
+}
+```
+
+## Automatic Usage Generation
+
+`goopt` automatically generates usage documentation based on defined commands and flags. To print the usage:
+
+```go
+parser.PrintUsage(os.Stdout)
+```
+
+To print the usage grouped by command:
+
+```go
+parser.PrintUsageWithGroups(os.Stdout)
+```
